Move features scripts from parsers to features_extractors
