--- edwar/parsers/acc/__init__.py
@@ -1,1 +1,0 @@
-from . acc import predict_activity, generate_features
