--- edwar/parsers/acc/models/finalized_raw_model_hand_selection.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3e84709f08e87c8083ddf80551cc088ca2ae858c8bf203870d0c13c2efa8f955
-size 241853817
