--- edwar/parsers/acc/models/finalized_raw_model_leftHand.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:174ad4d1bba6657b1eec7b046d2f28232d44d18bfdefdcaeda6cd5dbd22efe58
-size 268094729
