--- edwar/parsers/acc/models/finalized_raw_model_rightHand.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2104e126c1b49e4f2a6cc899a7d1caca0afb18e937bb02f9cc44537e7791a5ca
-size 273151289
