--- edwar/parsers/ibi/__init__.py
@@ -1,1 +1,0 @@
-from .ibi import check_ibi, calculate_hr, calculate_josue_hr
