--- edwar/parsers/stress/__init__.py
@@ -1,1 +1,0 @@
-from .stress import generate_data, predict_stress
